Address PR review: deduplicate BF helpers, remove dead code, simplify resimulate
- Extract shared helpers (path_has_conflict, reconstruct_path,
  extract_subgraph_edges) into bf_helpers.rs, used by both routing
  and pricing BF modules
- Remove dead amount_at method (moved to #[cfg(test)] only)
- Simplify resimulate_path: remove state-override tracking since
  forbid-revisits guarantees each pool appears at most once
- Fix doc comment to use proper module link instead of file reference
- Remove unused ProtocolSim import and dead simulation_amount accessor
