+++ fynd-core/src/algorithm/mod.rs
@@ -18,6 +18,7 @@
 //! 2. Implement the `Algorithm` trait
 //! 3. Register it in `registry.rs`
 
+pub(crate) mod bf_helpers;
 pub mod bellman_ford;
 pub mod bellman_ford_pricing;
 pub mod most_liquid;
